packaging works
Squashed commit of the following:

commit 742891d7a647447b9ae3dd6b58cf03cb2683b367
Author: Gambhiro <gambhiro.bhikkhu.85@gmail.com>
Date:   Mon Jan 31 18:14:25 2022 +0000

    upd

commit 4f4fa8058760c0a6db74783f7ba5cc11b40ccb46
Author: Gambhiro <gambhiro.bhikkhu.85@gmail.com>
Date:   Mon Jan 31 15:24:30 2022 +0000

    result item height

commit 3e72bc1af27bd1dc079964527e9e420da5bd58b0
Author: Gambhiro <gambhiro.bhikkhu.85@gmail.com>
Date:   Mon Jan 31 14:32:50 2022 +0000

    lookup actions

commit 3d988a021b8eb292e9ed4131265a5f968846cf3f
Author: Gambhiro <gambhiro.bhikkhu.85@gmail.com>
Date:   Sun Jan 30 19:57:53 2022 +0000

    bring lookup window to front

commit 7fc88fdabe429a974d6acc7a79fa545e838fdd9a
Author: Gambhiro <gambhiro.bhikkhu.85@gmail.com>
Date:   Sun Jan 30 19:53:55 2022 +0000

    button size

commit d126bb2597a46bd5ce2e3c51c64777db21b52984
Author: Gambhiro <gambhiro.bhikkhu.85@gmail.com>
Date:   Sun Jan 30 19:50:30 2022 +0000

    sutta and dict lookup in both windows

commit 87ba5dd2bb62cd4a519e92947ce9f042858c9656
Author: Gambhiro <gambhiro.bhikkhu.85@gmail.com>
Date:   Sat Jan 29 20:21:31 2022 +0000

    toggle pali keys

commit 92f3e50afd9201e4b7219dc3b579be97f4d747af
Author: Gambhiro <gambhiro.bhikkhu.85@gmail.com>
Date:   Sat Jan 29 11:49:58 2022 +0000

    upd

commit 073faf4cf66c7a15f7345431b7fb5dd5ad14142b
Author: Gambhiro <gambhiro.bhikkhu.85@gmail.com>
Date:   Sat Jan 29 09:35:06 2022 +0000

    upd

commit 28836275d51f87e6adb657fb04843af40a3d4bc3
Author: Gambhiro <gambhiro.bhikkhu.85@gmail.com>
Date:   Sat Jan 29 09:11:07 2022 +0000

    upd

commit a23b66b304bfb1e3b4b4897b8ca3ed83c25ff497
Author: Gambhiro <gambhiro.bhikkhu.85@gmail.com>
Date:   Fri Jan 28 20:32:22 2022 +0000

    upd

commit 4614f86110e5e3ba57331e42604822833f9c0a13
Author: Gambhiro <gambhiro.bhikkhu.85@gmail.com>
Date:   Fri Jan 28 20:07:32 2022 +0000

    upd

commit 48aa60767c69fdc88d57751928a12467becc40b6
Author: Gambhiro <gambhiro.bhikkhu.85@gmail.com>
Date:   Fri Jan 28 19:04:01 2022 +0000

    upd

commit 423039a36cc71ebda0b0f6cebe7d0cc4b00b4af7
Author: Gambhiro <gambhiro.bhikkhu.85@gmail.com>
Date:   Thu Jan 27 20:12:41 2022 +0000

    upd

commit 6cafca59a99bc9cdf09a4c3c55ea13b5a9f89a01
Author: Gambhiro <gambhiro.bhikkhu.85@gmail.com>
Date:   Thu Jan 27 19:47:38 2022 +0000

    upd

commit 4d9675d996d9cc10080b26f661f857b37a92725d
Author: Gambhiro <gambhiro.bhikkhu.85@gmail.com>
Date:   Thu Jan 27 19:23:21 2022 +0000

    upd

commit 0e3a5b9787eb733605ccbf70c6c002c71c0a27ea
Author: Gambhiro <gambhiro.bhikkhu.85@gmail.com>
Date:   Thu Jan 27 19:06:06 2022 +0000

    upd

commit e5b99f87a83da6ef825ec29921acc7da072b004d
Author: Gambhiro <gambhiro.bhikkhu.85@gmail.com>
Date:   Thu Jan 27 18:41:42 2022 +0000

    upd

commit 00652c6fe5e94a944ff54b68313a05164a3a822e
Author: Gambhiro <gambhiro.bhikkhu.85@gmail.com>
Date:   Thu Jan 27 17:53:42 2022 +0000

    upd

commit 0152d252f3cf98e1039c10df175145a9d383267c
Author: Gambhiro <gambhiro.bhikkhu.85@gmail.com>
Date:   Thu Jan 27 17:31:32 2022 +0000

    upd

commit e72ba3414532356254d526ed082b5308ebd6dcf9
Author: Gambhiro <gambhiro.bhikkhu.85@gmail.com>
Date:   Thu Jan 27 17:12:09 2022 +0000

    upd

commit 348a1b2155721d62178a28c5ada0a91966c52957
Author: Gambhiro <gambhiro.bhikkhu.85@gmail.com>
Date:   Thu Jan 27 16:48:27 2022 +0000

    upd

commit ba282ed8fcdc01f26f30b5949aa4a5c77e0b31c8
Author: Gambhiro <gambhiro.bhikkhu.85@gmail.com>
Date:   Thu Jan 27 16:02:41 2022 +0000

    upd

commit 5f9ba4e8bcaeda1d3245f1e23cb53077445550fa
Author: Gambhiro <gambhiro.bhikkhu.85@gmail.com>
Date:   Wed Jan 26 19:40:17 2022 +0000

    keys input

+++ simsapa/gui.py
@@ -4,6 +4,7 @@ import os
 import traceback
 import logging as _logging
 import logging.config
+from typing import Optional
 from PyQt5 import QtCore
 import yaml
 import threading
@@ -12,6 +13,8 @@ from PyQt5.QtGui import QIcon
 from PyQt5.QtWidgets import (QApplication, QSystemTrayIcon, QMenu, QAction)
 
 from simsapa import APP_DB_PATH, IS_LINUX, IS_MAC, IS_WINDOWS
+from simsapa.app.actions_manager import ActionsManager
+from simsapa.app.helpers import write_log
 from .app.types import AppData, create_app_dirs
 from .app.windows import AppWindows
 from .app.api import start_server, find_available_port
@@ -29,7 +32,9 @@ if os.path.exists("logging.yaml"):
 
 
 def excepthook(exc_type, exc_value, exc_tb):
+    write_log("excepthook()")
     tb = "".join(traceback.format_exception(exc_type, exc_value, exc_tb))
+    write_log(tb)
     logger.error("Error:\n", tb)
     w = ErrorMessageWindow(user_message=None, debug_info=tb)
     w.show()
@@ -38,14 +43,16 @@ def excepthook(exc_type, exc_value, exc_tb):
 sys.excepthook = excepthook
 
 
-def start(splash_proc: Popen):
+def start(splash_proc: Optional[Popen] = None):
     logger.info("start()")
+    write_log("start()", start_new=True)
 
     create_app_dirs()
 
     if not APP_DB_PATH.exists():
-        if splash_proc.poll() is None:
-            splash_proc.kill()
+        if splash_proc is not None:
+            if splash_proc.poll() is None:
+                splash_proc.kill()
 
         dl_app = QApplication(sys.argv)
         w = DownloadAppdataWindow()
@@ -56,24 +63,32 @@ def start(splash_proc: Popen):
 
     app = QApplication(sys.argv)
 
-    port = find_available_port()
-    daemon = threading.Thread(name='daemon_server',
-                              target=start_server,
-                              args=(port,))
-    daemon.setDaemon(True)
-    daemon.start()
+    try:
+        port = find_available_port()
+        write_log(f"Available port: {port}")
+        daemon = threading.Thread(name='daemon_server',
+                                target=start_server,
+                                args=(port,))
+        daemon.setDaemon(True)
+        daemon.start()
+    except Exception as e:
+        write_log(f"{e}")
+        # FIXME show error to user
+        port = 6789
+
+    actions_manager = ActionsManager(port)
 
     # FIXME errors on MacOS
     hotkeys_manager = None
 
     if IS_LINUX:
         from .app.hotkeys_manager_linux import HotkeysManagerLinux
-        hotkeys_manager = HotkeysManagerLinux(api_port=port)
+        hotkeys_manager = HotkeysManagerLinux(actions_manager)
     elif IS_WINDOWS:
         from .app.hotkeys_manager_windows_mac import HotkeysManagerWindowsMac
-        hotkeys_manager = HotkeysManagerWindowsMac(api_port=port)
+        hotkeys_manager = HotkeysManagerWindowsMac(actions_manager)
 
-    app_data = AppData(app_clipboard=app.clipboard(), api_port=port)
+    app_data = AppData(actions_manager=actions_manager, app_clipboard=app.clipboard(), api_port=port)
 
     app_windows = AppWindows(app, app_data, hotkeys_manager)
 
@@ -95,7 +110,7 @@ def start(splash_proc: Popen):
         menu.addAction(ac1)
 
         ac2 = QAction(QIcon(":dictionary"), "Lookup Clipboard in Dictionary")
-        ac2.setShortcut(_translate("Systray", "Ctrl+Shift+D"))
+        ac2.setShortcut(_translate("Systray", "Ctrl+Shift+G"))
         menu.addAction(ac2)
 
         if hotkeys_manager is not None:
@@ -116,13 +131,14 @@ def start(splash_proc: Popen):
 
     app_windows.show_update_message()
 
-    if splash_proc.poll() is None:
-        splash_proc.kill()
+    if splash_proc is not None:
+        if splash_proc.poll() is None:
+            splash_proc.kill()
 
     status = app.exec_()
 
     if hotkeys_manager is not None:
         hotkeys_manager.unregister_all_hotkeys()
 
-    logger.info(f"main() Exiting with status {status}.")
+    write_log(f"start() Exiting with status {status}.")
     sys.exit(status)
